# Conflicts:
#	drivers/flash/flash_stm32_qspi.c
#	drivers/gnss/gnss_quectel_lcx6g.c
#	drivers/i2c/i2c_ll_stm32.h
#	drivers/modem/modem_cellular.c
#	drivers/rtc/rtc_shell.c
#	drivers/wifi/esp32/src/esp_wifi_drv.c
#	include/zephyr/net/net_mgmt.h
#	include/zephyr/net/phy.h
#	include/zephyr/net/socket_net_mgmt.h
#	include/zephyr/net/wifi_mgmt.h
#	samples/net/secure_mqtt_sensor_actuator/src/main.c
#	subsys/mgmt/mcumgr/transport/src/smp_udp.c
#	subsys/net/ip/net_mgmt.c
#	subsys/net/l2/wifi/wifi_mgmt.c
#	subsys/net/lib/config/init.c
#	subsys/net/lib/shell/dns.c

+++ drivers/modem/modem_cellular.c
@@ -1733,7 +1733,7 @@ static int modem_cellular_pm_action(const struct device* dev, enum pm_device_act
 }
 #endif /* CONFIG_PM_DEVICE */
 
-static void net_mgmt_event_handler(struct net_mgmt_event_callback* cb, uint32_t mgmt_event,
+static void net_mgmt_event_handler(struct net_mgmt_event_callback* cb, uint64_t mgmt_event,
                                    struct net_if* iface) {
     struct modem_cellular_data* data =
         CONTAINER_OF(cb, struct modem_cellular_data, net_mgmt_event_callback);
